+++ lib/hexapdf/type.rb
@@ -73,6 +73,8 @@ module HexaPDF
     autoload(:IconFit, 'hexapdf/type/icon_fit')
     autoload(:AcroForm, 'hexapdf/type/acro_form')
     autoload(:Signature, 'hexapdf/type/signature')
+    autoload(:Outline, 'hexapdf/type/outline')
+    autoload(:OutlineItem, 'hexapdf/type/outline_item')
 
   end